Revert JDK-8241436: C2: Factor out C2-specific code from MacroAssembler
